aibackend/requirements.txt
fastapi
uvicorn
pydantic
transformers
torch
torchvision
Pillow
langdetect
keras
tensorflow
numpy